Created boolean_decomposition.py and regular_path_query.py
